requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.1
django-allauth==0.42.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-session-security==2.6.6
idna==2.9
jsonfield==3.1.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
python3-openid==3.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9